fix(contrib/clair): move commons function from old plugin package
