+++ sdk/plugin/common.go
@@ -6,7 +6,6 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
@@ -157,84 +156,6 @@ type Log struct {
 	Done               *timestamp.Timestamp `json:"done"`
 	StepOrder          int                  `json:"stepOrder"`
 	Value              string               `json:"val"`
-}
-
-// SendVulnerabilityReport call worker to send vulnerabiliry report to API
-func SendVulnerabilityReport(workerHTTPPort int32, report sdk.VulnerabilityWorkerReport) error {
-	if workerHTTPPort == 0 {
-		return nil
-	}
-
-	data, errD := json.Marshal(report)
-	if errD != nil {
-		e := fmt.Errorf("unable to marshal report: %v", errD)
-		Trace.Println(e)
-		return e
-	}
-
-	req, err := http.NewRequest("POST", fmt.Sprintf("http://127.0.0.1:%d/vulnerability", workerHTTPPort), bytes.NewReader(data))
-	if err != nil {
-		e := fmt.Errorf("send report to worker /vulnerability: %v", err)
-		Trace.Println(e)
-		return e
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		e := fmt.Errorf("cannot send report to worker /vulnerability: %v", err)
-		Trace.Println(e)
-		return e
-	}
-
-	if resp.StatusCode >= 300 {
-		e := fmt.Errorf("cannot send report to worker /vulnerability: HTTP %d", resp.StatusCode)
-		Trace.Println(e)
-		return e
-	}
-
-	return nil
-}
-
-// GetExternalServices call worker to get external service configuration
-func GetExternalServices(workerHTTPPort int32, serviceType string) (sdk.ExternalService, error) {
-	if workerHTTPPort == 0 {
-		return sdk.ExternalService{}, nil
-	}
-
-	req, err := http.NewRequest("GET", fmt.Sprintf("http://127.0.0.1:%d/services/%s", workerHTTPPort, serviceType), nil)
-	if err != nil {
-		e := fmt.Errorf("get service from worker /services: %v", err)
-		Trace.Println(e)
-		return sdk.ExternalService{}, e
-	}
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		e := fmt.Errorf("cannot get service from worker /services: %v", err)
-		Trace.Println(e)
-		return sdk.ExternalService{}, e
-	}
-
-	if resp.StatusCode >= 300 {
-		e := fmt.Errorf("cannot get services from worker /services: HTTP %d", resp.StatusCode)
-		Trace.Println(e)
-		return sdk.ExternalService{}, e
-	}
-
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		e := fmt.Errorf("cannot read body /services: %v", err)
-		Trace.Println(e)
-		return sdk.ExternalService{}, e
-	}
-
-	var serv sdk.ExternalService
-	if err := json.Unmarshal(b, &serv); err != nil {
-		e := fmt.Errorf("cannot unmarshal body /services: %v", err)
-		Trace.Println(e)
-		return serv, e
-	}
-	return serv, nil
 }
 
 //SendLog send logs to CDS engine for the current
